Bugfixes for install script and made uninstall script more robust

+++ dotfile_uninstall.sh
@@ -6,20 +6,37 @@ if ! [ -d ~/.original_dotfiles ]; then
 fi
 
 # Unlink Gitted dotfiles
-unlink ~/.bash_profile
-unlink ~/.bash_aliases
-unlink ~/.ssh/config
-unlink ~/.gitconfig
-unlink ~/.vimrc
+# ONLY unlink if they're symlinks in the first place
+[ -h ~/.bash_profile ] && unlink ~/.bash_profile
+[ -h ~/.bash_aliases ] && unlink ~/.bash_aliases
+[ -h ~/.ssh/config ] && unlink ~/.ssh/config
+[ -h ~/.gitconfig ] && unlink ~/.gitconfig
+[ -h ~/.vimrc ] && unlink ~/.vimrc
 
 # Copy over the backed up dotfiles
-mv ~/.original_dotfiles/.bashprofile ~/.bash_profile
+# Keep track of any failures to move
+FAILED_TO_MOVE=false
+mv ~/.original_dotfiles/.bash_profile ~/.bash_profile
+[ $? -ne 0 ] && FAILED_TO_MOVE=true
+
 mv ~/.original_dotfiles/.bash_aliases ~/.bash_aliases
+[ $? -ne 0 ] && FAILED_TO_MOVE=true
+
 mv ~/.original_dotfiles/.ssh/config ~/.ssh/config
+[ $? -ne 0 ] && FAILED_TO_MOVE=true
+
 mv ~/.original_dotfiles/.gitconfig ~/.gitconfig
+[ $? -ne 0 ] && FAILED_TO_MOVE=true
+
 mv ~/.original_dotfiles/.vimrc ~/.vimrc
+[ $? -ne 0 ] && FAILED_TO_MOVE=true
 
-rm -r ~/.original_dotfiles
+if [ $FAILED_TO_MOVE ]; then
+	echo "There was an error moving your original dotfiles back into place."
+	echo "Examine the output above and correct any mistakes, then delete .original_dotfiles when you are done."
+else
+	rm -r ~/.original_dotfiles
+	echo "Your original dotfiles have been replaced."
+fi
 
-echo "Dotfiles have been uninstalled."
-echo "Your original dotfiles have been replaced."
+echo "Git Dotfiles have been uninstalled."
